feat: add api-key option to login_required, simplify auth (remove temp cookies)
