+++ src/backend/app/auth/providers/google.py
@@ -25,7 +25,10 @@ from fastapi.exceptions import HTTPException
 from loguru import logger as log
 from requests_oauthlib import OAuth2Session
 
-from app.auth.auth_deps import create_jwt_tokens, set_cookies
+from app.auth.auth_logic import (
+    create_jwt_tokens,
+    set_cookies,
+)
 from app.config import settings
 from app.db.enums import HTTPStatus, UserRole
