+++ src/backend/app/projects/project_routes.py
@@ -46,7 +46,7 @@ from pg_nearest_city import AsyncNearestCity
 from psycopg import Connection
 from psycopg.rows import dict_row
 
-from app.auth.auth_deps import login_required, mapper_login_required, public_view
+from app.auth.auth_deps import login_required, public_endpoint
 from app.auth.auth_schemas import AuthUser, OrgUserDict, ProjectUserDict
 from app.auth.providers.osm import check_osm_user, init_osm_auth
 from app.auth.roles import check_access, mapper, org_admin, project_manager
@@ -148,7 +148,7 @@ async def read_project_summaries(
     db: Annotated[Connection, Depends(db_conn)],
     page: int = Query(1, ge=1),  # Default to page 1, must be greater than or equal to 1
     results_per_page: int = Query(13, le=100),
-    current_user: Annotated[Optional[AuthUser], Depends(public_view)] = None,
+    current_user: Annotated[Optional[AuthUser], Depends(public_endpoint)] = None,
     org_id: Optional[int] = None,
     user_sub: Optional[str] = None,
     hashtags: Optional[str] = None,
@@ -1304,7 +1304,7 @@ async def read_project(
 async def read_project_minimal(
     project_id: int,
     db: Annotated[Connection, Depends(db_conn)],
-    current_user: Annotated[AuthUser, Depends(mapper_login_required)],
+    current_user: Annotated[AuthUser, Depends(public_endpoint)],
 ):
     """Get a specific project by ID, with minimal metadata.
